Try to fix the grafana/k6 Docker image label from v0.40.0 to 0.40.0
This way it will match with previous releases. It was broken in https://github.com/grafana/k6/pull/2591 and the actual CI fix will happen in a separate bracnh and a proper PR. This is just a single commit we don't intend to merge in master, just to use the CI credentials to fix the issue for v0.40.0.