Fix TrendSink median calculation as part of Thresholds.Run
This commit addresses the issues outlined in #2755.

We assume that because the trend sink median computation can turn out
to be quite expensive (it involves sorting the array of values), it was
so far delegated to the last moment: thresholds evaluation and summary
generation. This detail appeared to be missed in the most recent
thresholds refactoring, and the `Calc` method was never used by the
`Thresholds.Run` method.

This commit ensures that in the same fashion as with Trend Sinks
percentiles, the thresholds evaluation also computes the sink's median
value when necessary.

This commit adds dedicated tests illustrating the behavior, and
demonstrating that it solves said issue.