Create deb822 sources file when creating the APT repo
This creates a deb822 sources file[1] that is uploaded to our dl.k6.io
S3 bucket the first time the script runs. This file allows us to simplify the k6
installation instructions on Debian-based Linux distributions.

So instead of users running gpg directly, dealing with any possible errors
from that[2], and specifying the legacy one-line list format manually, they only
need to download the sources file which embeds the signing key and
contains all information needed to install k6.

This format is supported since apt 1.1 (2015), and I tested it works in
Ubuntu 22.04 (apt 2.4.8).

Unfortunately, embedding the ASCII armored key is only supported since
apt 2.4, and recent versions of Debian have old apt versions (buster
has apt 1.8.2.3, and bullseye 2.2.4) where this is not supported.
It's likely that other Debian-based distros have newer apt versions, but
since it's not supported in Debian itself, we can't switch to this
format yet.

Using the deb822 format while still having to reference an
external keyring file doesn't provide much benefit over the current
legacy one-line format, so it's likely we'll use the current
instructions for years to come.

[1]: https://manpages.debian.org/buster/apt/sources.list.5.en.html#DEB822-STYLE_FORMAT

[2]: https://github.com/grafana/k6-docs/pull/1123